# docs/backend/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Geo URP API
  version: 1.0.0
  description: API base para RF-001 a RF-015
servers:
  - url: https://api.geo-urp.local/api/v1
paths:
...
  /public/events/calendar:
    get:
      summary: Calendario mensual de eventos
      tags: [Events]
      parameters:
        - in: query
          name: month
          required: true
          schema: { type: integer, minimum: 1, maximum: 12 }
        - in: query
          name: year
          required: true
          schema: { type: integer, minimum: 2000 }
      responses:
        '200': { description: OK }
  /admin/events:
    get:
      summary: Listar eventos (admin)
      tags: [Events]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
    post:
      summary: Crear evento
      tags: [Events]
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Creado }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT